Refactor naming scheme
