+++ image_viewer/util/_os_nt.pyi
